Consolidate tests, condense file names
