+++ alea/discrete/__init__.py
@@ -1,5 +1,5 @@
-from .discrete_randvar import *
+from .randvar import *
 from .root_randvar import RootDiscreteRandVar
 from .function_randvar import *
 from .special_randvar import *
-from .discrete_randvec import *
+from .randvec import *
