+++ alea/discrete/function_randvar.py
@@ -1,4 +1,4 @@
-from .discrete_randvar import DiscreteRandVar
+from .randvar import DiscreteRandVar
 
 import copy
 import itertools
