+++ alea/discrete/randvar.py
