+++ tests/test_discrete_randvec.py
